SUN-76: Refactor step definition classes
